batch7 high-effort contract-spec review fixes (5 P2, all verified): T95 now scales the token-window column offset by req_to_token.stride(1) and reads the four per-request metadata vectors through their own strides; T98 scales both half loads by kv_buffer.stride(1) and reads loc through loc.stride(0); T96 reads indices through indices.stride(0) and normalizes non-contiguous q/k/v with explicit .contiguous() copies (PR#69 precedent) instead of asserting; T97 promotes all offset/stride arithmetic to i64 including the flat-path offs computation (pid*BLOCK itself overflows past 2^31 elements); regressions added for column-sliced req_to_token, column-sliced kv_buffer, sliced indices, and transposed q/k/v

+++ src/flaggems_sglang/ops/fused_sigmoid_mul.py
@@ -28,11 +28,12 @@ def _fused_sigmoid_mul_kernel(
     GATE_CONT: tl.constexpr,
     BLOCK: tl.constexpr,
 ):
-    offs = tl.program_id(0) * BLOCK + tl.arange(0, BLOCK)
+    offs = tl.program_id(0).to(tl.int64) * BLOCK + tl.arange(0, BLOCK).to(tl.int64)
     mask = offs < numel
     a_ptr = attn + offs
     g_ptr = gate + offs
     if not (ATTN_CONT and GATE_CONT):
+        # stride math in i64: large strides or numel can overflow i32
         row = offs // hidden
         rem = offs - row * hidden
         if not ATTN_CONT:
